ActividadPolicy

+++ resources/views/eventoActividad/index.blade.php
@@ -27,10 +27,13 @@
           <td> {{$value->hora_inicio}}</td>
           <td> {{$value->hora_fin}}</td>
           <td>
-            {{ Form::open(['method' => 'DELETE','route' => ['actividad.destroy', $value->id],'style'=>'display:inline'])}}
-            {{ Form::submit('Eliminar')}}
-            {{ Form::close()}}
             <a href="{{route('evento.actividad.show',[$id_evento,$value->id])}}"> Mostrar</a>
+            @can('modify',$value)
+              <a href="{{route('evento.actividad.edit',[$id_evento,$value->id])}}"> Editar</a>
+              {{ Form::open(['method' => 'DELETE','route' => ['actividad.destroy', $value->id],'style'=>'display:inline'])}}
+              {{ Form::submit('Eliminar')}}
+              {{ Form::close()}}
+            @endcan
           </td>
         </tr>
         @endforeach
